tests/FFI: add test code from redhat/edge/base-image repo
It makes sense to have that here, to be able to execute against qm
directly.

This is a plain integration of the source files, rougly categorized. The
Makefile scheme was followed, but it will be updated as part of
a separate pull request to handle more files more efficiently.

The way the Makefile works now it was easiest to put all target as
dependencies of the 'all' target.

Integration with the test setup will be a follow-up.